XQuartz: 64bit fix for screen origin in AppleWM (cherry picked from commit 60a757d2802a5c34acd91ca9a052937b5a169ede)

+++ hw/xquartz/applewm.c
@@ -137,7 +137,7 @@ AppleWMSetScreenOrigin(
     WindowPtr pWin
 )
 {
-    long data[2];
+    int32_t data[2];
 
     data[0] = (dixScreenOrigins[pWin->drawable.pScreen->myNum].x
                 + darwinMainScreenX);
